stage-7: fix codeclimate 3
