+++ src/main/java/ru/vk/itmo/test/viktorkorotkikh/LZ4CompressedLSMServiceFactoryImpl.java
@@ -1,0 +1,17 @@
+package ru.vk.itmo.test.viktorkorotkikh;
+
+import ru.vk.itmo.Service;
+import ru.vk.itmo.ServiceConfig;
+import ru.vk.itmo.dao.Config;
+import ru.vk.itmo.test.ServiceFactory;
+
+@ServiceFactory(stage = 7)
+public class LZ4CompressedLSMServiceFactoryImpl implements ServiceFactory.Factory {
+    @Override
+    public Service create(ServiceConfig config) {
+        return new LSMServiceImpl(
+                config,
+                new Config.CompressionConfig(true, Config.CompressionConfig.Compressor.LZ4, 4096)
+        );
+    }
+}
