+++ src/main/java/ru/vk/itmo/test/viktorkorotkikh/reports/stage7/report.md
@@ -251,6 +251,8 @@ wrk смог пропихнуть только 1.5k RPS. Нужно снижат
 100.000%   38.44s 
 ```
 
+При этом уже 37 запросов завершилось с таймаутом.
+
 Точка разладки ещё сильнее сместилась:
 
 [ZSTD-4096-GET-200.txt](ZSTD-4096-GET-200.txt)
